requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
distro==1.9.0
greenlet==3.3.2
groq==1.1.2
h11==0.16.0
hdbscan==0.8.42
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
pgvector==0.4.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
scikit-learn==1.8.0
sentence-transformers==5.3.0
slack==0.0.2
slack_bolt==1.27.0
slack_sdk==3.41.0
sniffio==1.3.1
SQLAlchemy==2.0.48
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.11